// shared/types/NFTMarketplaceV2.ts
export declare namespace NFTMarketplaceV2 {
  export type ListedTokenStruct = {
    tokenId: BigNumberish;
    owner: AddressLike;
    creator: AddressLike;
    price: BigNumberish;
    currentlyListed: boolean;
    royalty: BigNumberish;
  };

  export type ListedTokenStructOutput = [
    tokenId: bigint,
    owner: string,
    creator: string,
    price: bigint,
    currentlyListed: boolean,
    royalty: bigint
  ] & {
    tokenId: bigint;
    owner: string;
    creator: string;
    price: bigint;
    currentlyListed: boolean;
    royalty: bigint;
  };

  export type SaleHistoryStruct = {
    tokenId: BigNumberish;
    previousOwner: AddressLike;
    newOwner: AddressLike;
    salePrice: BigNumberish;
    saleTimestamp: BigNumberish;
  };

  export type SaleHistoryStructOutput = [
    tokenId: bigint,
    previousOwner: string,
    newOwner: string,
    salePrice: bigint,
    saleTimestamp: bigint
  ] & {
    tokenId: bigint;
    previousOwner: string;
    newOwner: string;
    salePrice: bigint;
    saleTimestamp: bigint;
  };
}